+++ bvm/ballerina-core/src/main/java/org/ballerinalang/model/values/BInteger.java
@@ -40,7 +40,7 @@ public final class BInteger extends BValueType implements BRefType<Long> {
 
     @Override
     public int byteValue() {
-        return (byte) this.value;
+        return (int) this.value;
     }
 
     @Override
