+++ bvm/ballerina-core/src/main/java/org/ballerinalang/util/codegen/cpentries/ByteCPEntry.java
@@ -25,13 +25,13 @@ package org.ballerinalang.util.codegen.cpentries;
  */
 public class ByteCPEntry implements ConstantPoolEntry {
 
-    private int value;
+    private byte value;
 
-    public ByteCPEntry(int value) {
+    public ByteCPEntry(byte value) {
         this.value = value;
     }
 
-    public int getValue() {
+    public byte getValue() {
         return value;
     }
